feat: relocate single-service components to feature modules
Move components from components/shared/ to their respective feature
module component directories:

- cel-editor, starlark-editor, saga-timeline → features/sagas/components/
- quality-ladder-badge → features/positions/components/
- create-valuation-feature-dialog → features/reference-data/components/

Create index.ts barrel files in each feature's components/ directory.
Update all import paths in consuming files.
Remove moved exports from components/shared/index.ts.

Part of 034-frontend-alignment.3
