--- frontend/src/components/shared/index.ts
@@ -1,7 +1,0 @@
-export { SagaTimeline } from './saga-timeline';
-export type { SagaTimelineProps, SagaStep } from './saga-timeline';
-
-export { QualityLadderBadge } from './quality-ladder-badge';
-
-export { CreateValuationFeatureDialog } from './create-valuation-feature-dialog';
-export type { CreateValuationFeatureDialogProps, AccountType } from './create-valuation-feature-dialog';
