+++ frontend/src/features/positions/components/index.ts
@@ -1,0 +1,1 @@
+export { QualityLadderBadge } from './quality-ladder-badge';
