+++ frontend/src/features/positions/components/quality-ladder-badge.test.tsx
