+++ frontend/src/features/reference-data/components/index.ts
@@ -1,0 +1,2 @@
+export { CreateValuationFeatureDialog } from './create-valuation-feature-dialog';
+export type { AccountType, CreateValuationFeatureDialogProps } from './create-valuation-feature-dialog';
