+++ frontend/src/pages/payments/payment-detail.tsx
@@ -8,7 +8,7 @@ import { Button } from '@/components/ui/button'
 import { MoneyDisplay } from '@/shared/money-display'
 import { StatusBadge } from '@/shared/status-badge'
 import { TimeDisplay } from '@/shared/time-display'
-import { SagaTimeline } from '@/components/shared/saga-timeline'
+import { SagaTimeline } from '@/features/sagas/components/saga-timeline'
 import { AuditTrail } from '@/shared/audit-trail'
 import { EntityLink, Breadcrumbs } from '@/shared'
 import { useTenantSlug } from '@/hooks/use-tenant-context'
