+++ frontend/src/pages/positions/detail.tsx
@@ -5,7 +5,7 @@ import { Card } from '@/components/ui/card'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
 import { MoneyDisplay } from '@/shared/money-display'
 import { TimeDisplay } from '@/shared/time-display'
-import { QualityLadderBadge } from '@/components/shared/quality-ladder-badge'
+import { QualityLadderBadge } from '@/features/positions/components/quality-ladder-badge'
 import { DirectionBadge } from '@/shared/direction-badge'
 import { EntityLink, Breadcrumbs } from '@/shared'
 import { Skeleton } from '@/components/ui/skeleton'
