+++ frontend/src/pages/starlark/detail.tsx
@@ -5,7 +5,7 @@ import { Card } from '@/components/ui/card'
 import { Button } from '@/components/ui/button'
 import { StatusBadge } from '@/shared/status-badge'
 import { TimeDisplay } from '@/shared/time-display'
-import { StarlarkEditor, type ValidationError, type ComplexityMetrics } from '@/components/shared/starlark-editor'
+import { StarlarkEditor, type ValidationError, type ComplexityMetrics } from '@/features/sagas/components/starlark-editor'
 import { Breadcrumbs } from '@/shared'
 import { useApiClients } from '@/api/context'
 import { SagaStatus, ErrorCategory } from '@/api/gen/meridian/saga/v1/saga_registry_pb'
